ShipmentProvidersVC: Use WooBasicTableViewCell
The selection (checkmark) behavior of the previously used `StatusListTableViewCell` requires us to preselect the item using `table.select()`. This seems like an overhead since it requires us to search for the correct `indexPath`, possible `O(n)`, whenever the list is reloaded.

I believe it’s better if the setting of the checkmark is more reactive — do it in `cellForRowAt`. Then we also don’t need to worry about _when_ the table reloading happens.