Split screenshot lanes up to run in CI
This PR splits up the phaess of screenshot generation in order to be able to run in parellel:

**Build Screenshots**
Builds the WooCommerce app, as well as the ScreenshotRunner app to drive the UI

** Take Screenshots**
Launches the WooCommerce app and runs the ScreenshotRunner app in the background to take the screenshot. Stores them in the `screenshots` directory.

** Rebuild Screenshot Devices **
Adds support for rebuilding screenshot devices – this is copied from Android and helps ensure that the devices we need for generating screenshots are always present.

** Create Screenshot Summary **
Adds support for crawling the `screenshots` directory tree and assembling a nice HTML file with references to all of the images. This can be run on its own following the parallel generation of the each screenshot locale in order to put everything back together afterward.

** Screenshots **
Puts everything together – it’ll run the entire screenshot process on a local machine by calling all of the lanes, as appropriate.

There are several new helpers in order to avoid a lot of code duplication.