Add OrderAction.fetchFilteredAndAllOrders
This action:

- Fetches the first page of the given filter
- AND fetches the first page of all orders (any filter)

This will be used to solve the problem of users not being able to see the recent
changes on the Processing tab when performing this scenario:

1. Load the Processing tab.
2. Change one of the processing orders to "Completed" in the web.
3. Pull to refresh on the Processing tab.

They would not be able to see the up to date _completed_ order because the
fetch request will only include the "processing" filter
(`GET /orders?filter=processing`).

Fetching both "processing" and all orders lists should retrieve the most up to
date version of the order. We can also augment this by passing `true` for the
`deleteAllBeforeSaving` argument so that we are sure that the user is viewing
the most up to date list of orders.