Orders: Delete and perform dual-fetch on pull-to-refresh
If the user uses pull-to-refresh on the Processing tab, we should delete
all the orders so we are sure that the user will be viewing the most recent
orders.

We perform a dual fetch since there are 2 tabs and it's a better experience if
the user do not have to wait for one of the tabs to be loaded.

The dual-fetch also happens on `viewWillAppear` but we don't delete the orders
during this time since the user may just be switching between different tabs and
we don't want them to load the "next" pages again.