MockupNetwork: Fix dequeueing when using `useResponseQueue`
When dequeuing a filename, we search for the `Queue` struct in `responseQueue`
and call `dequeue()` on it if it exists. The problem with this is that we are
temporarily assigning the `Queue` to `var queue`. Swift structs are passed by
value so we're basically **dequeuing on a copy**.

Directly accessing the `Queue` through the `responseQueue` will make sure that
we're operating on the same `Queue` instance and not a copy.