Use --repo-update flag when running pod install from Rake
If CocoaPods' local specs repo out-of-date, `pod install` might not find
some of the specified pods and fail.

The `--repo-update` flag make CocoaPods always update its local specs
repo, so that that kind of failure won't happen.

This does add a few seconds to the task, but I feel it is time well spent
because it will save a lot more time troubleshooting a failure in `rake
dependencies` which might not be obvious to someone not used to dealing
with CocoaPods.