Make loadOrCreateGeneralAppSettings self-recover
This fixes the issue where we can't save anything if the file does not exist.
Now, the function will just return an empty `GeneralAppSettings`. This was my
original intent. I just failed to see it last time because `MockInMemoryStore`
hid this problem. ¯\_(ツ)_/¯

This also fixes the failing test in AppSettingsStoreTests.