ReviewDetailsVC: Remove handling of entityListener.onDelete
This causes an issue when handling push notifications. Consider this scenario:

1. Delete the app to clear the data.
2. Run and login. Do not load the Reviews list.
3. Forcefully terminate the app.
4. Send a Review push notification.
5. Tap on the push notification.
6. The Review Details page will be shown but it will be immediately dismissed.

This happens because the Reviews list page is being loaded for the first time
and it will delete all the reviews. Thereby, causing ReviewDetailsViewController
to dismiss itself.

It looks like it's fine to keep the ReviewDetailsViewController shown even if
the underlying ProductReview data has been deleted in the database. I can still
perform the Spam or Delete actions like usual.