Fix discrepancies in order details view in between the moment a refund is issued and the moment the order is refreshed with data from the API
The issue is that we were always expecting refunds to have a negative value, which is not true. There is a case when a refund can have a positive value: in between the refund is issued in the app and the moment we update the order from the API.

The fix touches the way refunds are calculated and the way they are presented. I don't love this solution, but tried to submit a refund with a negative value didn't work (the API creates the refnd but sets it to zero) and forcing all refunds to be negative at the Store seemed like could be more dangerous.