Coupon Model
Adds the Coupon model according to the API docs, except for a few fields which would add confusion and are unlikely to be required, e.g. the dates in the site’s timezone.

Synthesised codable conformance is used as far as possible, which gives us more compiler-provided safety, e.g. around key decoding correctness.