CouponsRemote with loadAllCoupons method
The Remote is closely coupled with the Mapper here, which I’d like to improve for testability.

The implementation here follows the existing pattern, but some further improvement to allow injection of the mapper would be helpful, so that we could test for what the siteID passed to the mapper is independently of the results of parsing coupons.

CouponsRemote with loadAllCoupons method

The Remote is closely coupled with the Mapper here, which I’d like to improve for testability.

The implementation here follows the existing pattern, but some further improvement to allow injection of the mapper would be helpful, so that we could test for what the siteID passed to the mapper is independently of the results of parsing coupons.