Update WeightFormatter to use MeasurementFormatter
The current implementation basically joins the value and unit with a space.
This works in many cases, but not for all locales. RTL languages will need the
components reversed, some languages (like Chinese) prefer omitting the space
between the number and the unit, and many others translate the unit.

By using MeasurementFormatter when possible, we get all these localizations for free.
When the weight unit isn't recognized, we fall back to the previous behavior.

One noticeable difference is that Woo seems to use `lbs` as the abbreviation
for pounds, while iOS uses `lb`.