CouponList: handle pagination in the view model
Originally, I intended to use the `PaginatedListViewControllerStateCoordinator` to handle pagination state, however, I want to ensure that this is tested and part of the view model.

Using the state coordinator directly is possible, but leads to a lot of translation of states. In particular, the ability to distinguish between the various loading states (.loading, .loadingNextPage, .refreshing) is difficult with the existing class.

In this change, I’ve instead taken inspiration from the PaginatedListViewControllerStateCoordinator, with the intention to later port these changes back to that class, so that other screens in the app could also benefit from simplified loading state handling.