Allow host app to override login error UI in GetStartedViewController
This reverts some of the changes in 0271ba0. That commit removed the ability for
the host app to override errors.

I'm applying it a bit differently here though. There was a condition added in
d0b6ce4 which automatically added an error dialog if the
`WordPressAuthenticator.shared.configuration.enableSignUp` is `false`. The
`sendEmail` is no longer called so I'm adding the override directly in `handleLoginError()`.