Define User model in Networking
The User model represents "local accounts" on specific sites, and will generally be populated through the `/wp/v2/sites/$site_id/users/me` endpoint. The important difference with the regular `/me` endpoint, is that it returns a list of `roles` for the User on the site. The role field is required to determine whether the user can access the shop through the WooCommerce app or not.