Implement role error flow in AppCoordinator
If the user is authenticated and has a default store, AppCoordinator will now check whether there are previous eligibility errors. If so, then the user will be presented with the role error page.

Note that in the case of unexpected errors (i.e. connection loss), the logic will stick with whatever result was stored locally. In the case that the user has been granted the correct roles but got stuck in the error page – there is a retry button to re-check user eligibility.