Prevents crash when reconnecting to a reader
This was a hard one to track, and I don't understand all the moving pieces, but the essence of it is that, when we use `htmlToAttributedString` to create the attributed string for the "Learn more" footer, this will internally load webkit. While it's loading, the NSAttributedString internals will return control of the execution to the run loop, which might start running async blocks of code.

In this particular case, this can lead to situations where the table view is recreating its cells, and a callback to one of the StripeTerminal methods gets executed and attempts to update the UI. I didn't find the specific problem, but we would likely have several instances where there was a UI callback while the UI was in an inconsistent state.

This commit doesn't solve the issue perfectly, as it would still happen the first time the cell is rendered, but at least any subsequent calls will be cached. At least this fixes the crash, although we might want to reconsider this approach in the future, and maybe use something simpler like markdown for links in translations, or process them at build time.