Remove unused Buddybuild prebuild script
I'm not sure when we stopped using Buddybuild, but it certainly unused
at the moment. On top of that, Apple shut it down, so, even if we wanted
to keep this "just in case, for the future" there would be no point.

I also checked with `git grep buddy` and found no match.