KeyboardStateProvider: Use *will* instead of *did* events
Using the `did*` events causes incorrect state values when classes like
`KeyboardFrameObserver` tries to get the current state. `KeyboardFrameObserver`,
in particular`, observes `will*` events.

For example, what can happen is that `KeyboardFrameObserver` will receive a keyboard
_hidden_ event first but then `KeyboardStateProvider` still thinks that the
keyboard is visible because it is observing `didHide*`.

I'm aligning the events that are being used so that these classes work well together.