Don't request line_items in the order list
In the order list view, we don't show details about the individual
products in the order. Requesting full details about all the products
can have a significant impact on performance.

Here we differentiate between fields we need for single orders and a
list of orders. For now the only difference is that single orders do
load the line items.