Prevent order items from updating if items is empty
When we go from the single order screen to the list of orders, we update
the order list. Since the oder list request no longer includes line
items, updating the order items ends up deleting them all. Then when we
go back to the single order screen, it has to fill in the list again.
Going back and forth between the order list and a given order will cause
it to have to fill in the list of orders every time.

If the order items is empty, it's a pretty good sign that we're on the
list. However, it would be better to allow the order items to be nil and
then we could look for `readOnlyOrder.items != nil`, which would be a
better indication that the data just doesn't include line items.