3722 Fix crash on accessibilityLabel for media
A rare crash when a user opened the MediaPicker-iOS was due to unchecked access of the `filename` property on Media.

`filename` is an optional part of the protocol, so this crash will also need to be fixed in MediaPicker-iOS, as the asset class should be checked for conformance before the optional property is accesed. See https://github.com/wordpress-mobile/MediaPicker-iOS/issues/371 for progress.

MediaPicker-iOS attempts to use the `filename` to create an accessibility label for a document, which will be of the form "Document: {filename}". This crash is very rare (4 incidents in the last month) likely because everywhere we use MediaPicker-iOS, we filter to show images only, which use the date in the accessibility label; for images, filename is not accessed.

This change adds the filename to the fields we request on the MediaRemote, and implements it on the WPMediaAsset protocol. Similarly, `fileExtension` is an optional conformance on the same protocol which we could trivially add, so that is included here too.

These changes will make this crash impossible within the WCiOS app, and slightly improve our users VoiceOver experience, but we'll also fix the library.