Model PaymentIntent metadata matching Stripe
Previously, for any PaymentIntent metadata created using our convenience initializers, we stored the orderID as an Int64, however Stripe would transform this to a String. This is due to Stripe sending the metadata over the network in a x-www-urlformencoded request, without any type information, and returning it as a String.

This meant that the type stored in the [AnyHashable: Any] for orderID could be a String, or an Int64, depending how the PaymentIntent was created (in tests with the convenience intitializer, vs from Stripe's network response.)

This change maintains the Int64 for a convenient(!) convenience init, while also matching the real-life Stripe metadata for accuracy in tests.