5088 Retry reader connection from scratch
Previously, we continued the search when the user tapped `Try Again`. Unfortunately, this would never find the reader they had just failed to connect to, so it's not really trying again.

This change resets the search and re-runs it, so the original reader can be found again. There may be better ways to do this.