Use cocoapods-catalyst-support to exclude Terminal
This does not build on Mac Catalyst yet, with lots of errors `No such module 'StripeTerminal'`. That's because we need to exclude its use with `#if targetEnvironment(macCatalyst)` compilation directives, but doing so leads to further issues at the call sites, so I've left that out of here to avoid confusion.

Run a post install phase to workaround signing issues