Replace `fastlane-plugin-test_center` with vanilla `run_tests` call
Xcode 13.0 introduced "Test Repetition Mode" in the `.xctestplan`
configuration, which we can use to replicate the `try_count` feature
provided by the plugin.

Dropping `fastlane-plugin-test_center` is necessary because of a bug
between it and recent version of Fastlane.

Going back to vanilla `run_tests` also required changing the way the
generate the XML test report, because of the difference in return
parameters with `multi_scan`. The good news is that the resulting setup
is simplified given that, either at the `run_tests` or Xcode 13 level,
we now have all the test results in a single `.xctestresult` file.

While touching the report generation code, I decided to not remove the
reports generated by Fastlane in HTML and JUnit formats: storage is
cheap. Besides, we spin-up new VMs on every CI job :)

Notice also that I did not enable test repetition in the unit tests
configuration, despite them having `try_count: 3` when called via
Fastlane in CI. It's okay to use test repetition in UI tests to account
for their intrinsic flaky nature. However, unit tests should not be
flaky and therefore we shouldn't swallow any kind of failure, but
rather notify them and address them ASAP.

See
https://github.com/lyndsey-ferguson/fastlane-plugin-test_center/issues/358

See also an example of failing build without this change:
https://app.circleci.com/pipelines/github/woocommerce/woocommerce-ios/17336/workflows/7684b230-2c15-4e57-b554-270035ac821f/jobs/33462

Add note about `fail_build: false` usage in `Fastfile`

WIP