6278 Add localization for ReaderDisplayMessage
Stripe provide instructional messages to help merchants and their customers complete a card transaction, e.g. “Swipe card”, “Try another read method”, “Multiple contactless cards detected”, “Remove card”.

Stripe do not provide localizations for these messages, they are provided as an enum, and a method which converts that enum to an English string.

This change copies the current output of Stripe’s `Terminal.stringFromReaderDisplayMessage` for use as the English text when these messages are displayed to the user, but also uses it as the key for localization.

There is no change to what’s displayed for merchants using English, likely the vast majority of our IPP merchants.