6357 Accept IPP for confirmed bookings
Bookings which have been confirmed by an admin have a specific payment method name on the order, `wc-booking-gateway`.

This meant that they could not be paid for with a card reader, as the only payment methods supported were `cod, woocommerce-payments, none`.

This adds the `booking` payment type to the supported payment methods eligible for IPP.