6461 Don’t say `Remove card` when starting refund
The refund flow works well even when the customer puts their card in too early. A new transaction won’t be permitted by the reader if it’s not taken out after the last one, so there’s low risk that the refund will be done to a previous customer’s card.