6643 Allow retries of failed interac refunds
When an interac refund fails, we show an error with a “Try again” button.

When the user taps that, we attempt to cancel the Interac refund “intent” using the `refundCancelable` that Stripe Terminal gives us in response to `SCPTerminal.collectRefundPaymentMethod()`

When there’s been a failure or error in the refund, the `refundCancelable` that Stripe gave us ends up in an unexpected situation: it’s not `completed`, but it also doesn’t call back the completion handler when we call `cancel`. This meant that our previous “Try again” handler did not start the new refund attempt.

Cancelling a refund after Stripe have called the completion handler from `processRefund`, or given us an error from `collectRefundPaymentMethod`, shouldn’t be required: those calls from Stripe have already told us that the original refund attempt is complete (even if failed), so we can’t cancel it anyway.

This change sets the `refundCancelable` to nil when Stripe have told us that the refund is complete, resulting in a `noRefundInProgress` error being returned to the `handleRefundFailureAndRetryRefund` handler. We don’t need to worry about this error or whether the cancellation was successful: in either case, we should try the refund again.

This change does not address that some refund errors should not be retriable: that will follow in future work.