Always use dedicated lane for codebase localization update
This lane is almost redundant, given it only calls an action. But, it's
handy to have it because you can have tab completion in the terminal for
lanes, but not for running actions. That is, if we have a dedicated lane
it's easier to run and discover in the terminal than a dedicated action.