Move `Main` and `LaunchScreen` storyboards outside of `Base.lproj`
The new location is `WooCommerce/Classes/ViewRelated/` for _both_.
Previously, `Main.storyboard` was there, but `LaunchScreen.storyboard`
was in the `Resources/` folder. That seemed inconsistent in particular
after the move away from the `Base.lproj` approach in this commit.

Xcode uses `Base.lproj` as the folder for the "development language".
From the docs:

> The `Base.lproj` folder contains all the `.storyboard` or `.xib` files in
> the development language. The folders for languages that you add to your
> project contain a strings file for each `.storyboard` or `.xib` file in your
> project. The development language folder doesn’t contain strings files
> for the `.storyboard` and `.xib` files because they don’t require
> translation in the development language. All the language folders
> contain a `InfoPlist.strings` file used to localize bundle properties,
> such as the app name. Any other localized resources—such as strings
> files that you generate and use from your code—appear in these
> language folders.
>
> For example, for a universal iOS app that uses English as the
> development language, the `Base.lproj` folder contains
> `Main_iPad.storyboard` and `Main.iPhone.storyboard,` and the `en.lproj`
> folder contains `InfoPlist.strings.` If you add the German language, Xcode
> creates a `de.lproj` folder containing a `InfoPlist.strings,`
> `Main_iPad.strings,` and `Main_iPhone.strings` file. The files in the
> `de.lproj` folder contain placeholder text that needs to be translated to
> German.

The way I read that documentation is that we should have files in the
`.storyboard` and `.xib` files in `Base.lproj` only if we plan to
localize them. But in our case `LaunchScreen.storyboard` only shows the
Woo logo image and `Main.storyboard` is an empty container. That is, we
can safely move them away from `Base.lproj` and not have that affect the
app.

The reason I noticed this is that the new localization automation cannot
cope with an `.lproj` folder (in this case, `Base.lproj`) where it can't
find a `Localizable.strings` file.

One alternative option would be to update our tooling to selectively
skip `Base.lproj`, since the Apple docs state it shouldn't contain
`.strings` files.

That's an option, but in the meantime, this change would allow us to
move on with the current localization tooling improvements.