6607 Check onboarding state before collect payment
On opening Order Details, this checks the onboarding state for In-Person-Payments.

When the merchant taps `Collect Payment`, it shows onboarding if required.

If onboarding is displayed, when the state becomes `.complete`, the onboarding flow is dismissed and the card payment process started.

There is code added which lays the groundwork for checking the reader connection status first, but at present, only the onboarding state is used.