6607 Explicit refresh for InPersonPaymentsVM
The InPersonPaymentsViewModel caused a refresh of its `CardPresentPaymentsOnboardingUseCase` when it was created.

With bac5a7412bf479c682565057eeb31b04fc39793c, we delayed the call to `refresh` until a card reader was connected. Since we share the useCase between these two classes, that change was ineffective, because the use case had already been refreshed.

This updates the InPersonPaymentsViewModel to require an explicit refresh, so that it can use an existing useCase without refreshing when appropriate.