6753 Pop to SwiftUI in Simple Payments flow
The Simple Payments flow provides a UINavigationController as the root controller for presenting card present payments UI.

We also use this for popping the onboarding view controller, when we’ve presented it and no longer need it because onboarding is complete.

This changes our dismiss code to work with a viewController which either _is_ a nav controller, or which _has_ a nav controller.