6807 Perform refund onboarding check earlier
Previously, the `Connecting to your account` onboarding screen would show for all attempts to perform a card present refund (i.e. an Interac refund).

This was due to the CardPresentPaymentsOnboardingPresenter being created then used in the submission step.

This change moves the creation of the presenter to the confirmation step, so that the readiness check has been completed by the time the merchant confirms the refund.

When the merchant confirms the refund, most of the time there will be no onboarding issues, so now they won’t see the onboarding flow at all.