6607 Fix unexpected Order Details reader connection
Previously, it was possible to unexpectedly trigger a reader connection by tapping `Connect Payment`, seeing an onboarding state, going back to Order Details, fixing the onboarding state, and then pulling to refresh.

The `readinessSubscription` does this when you resolve an onboarding issue, and that’s deliberate, but it’s meant to happen while you’re on the Onboarding screen and you tap a “Refresh after Updating” or similar button in an onboarding screen.

When you’re back on the Order Details screen, we shouldn’t remain subscribed to the readiness state.