6766 Replace calls to checkCardReaderConnected
publishCardReaderConnections is a more flexible option introduced for the in-payment Onboarding flow checks, which does not rely on a `sink(recieveCompletion:)` call.

These two actions are so similar, that it doesn’t make sense to keep both implementations around. This commit refactors existing calls to `checkCardReaderConnected` to use `publishCardReaderConnections` instead.