Fix warning in addCloseNavigationBarButton helper
The `addCloseNavigationBarButton` helper function used `self` as a default value for the `target` parameter.

From Xcode 13.3, this produced a warning:

> `'self' refers to the method 'UIViewController.self', which may be unexpected`

That was due to the function signature being available before `init` is complete, at which time `self` does not refer to self. The compiler-suggested fix (expressly specifying `UIViewController.self`) causes a crash when the close button is tapped on any view controller which uses this helper.