Delete unused `fastlane/metadata/default/promo_screenshot_*.txt` files
`fastlane/metadata/default/` should be the source of truth for all
metadata, but we haven't yet updated the part of the logic that has to
do with the sceenshots localization and generation.

Until we do that, having those file in there is confusing and
wasteful, because they are never read as the source of truth for what
they stand for, which is instead in
`fastlane/appstoreres/metadata/source`.

This is a followup to the work in 560b1dd7bb17c0d5a2484bd0c5e7b9fc3352a97c

See discussion at
https://github.com/woocommerce/woocommerce-ios/pull/6936#discussion_r886724474