6780 Prevent flickering of Collect Payment
We have decided to remove the check for whether there is a CPP eligible account or not when we show the Collect Payment button.

The check will now be handled by the onboarding flow, and removing the reliance on this results controller prevents the flickering we used to see on this button.