Refactored BottomSheet disclosureIndicator
Previously the disclosureIndicator had to be shown if you want a selection state on tapping a button in the bottom sheet.

This change extracts the disclosure indicator logic from isActionable

Now the user can see the cell being tap with a selection state despite not having a disclosure indicator.

Not all cells actionable cells display more information so not all need to display a disclosure indicator.