6782 Rename payments flow analytics with generics
The Simple Payments Payment Methods screen is suitable for use as a payment method choice from the Order Details screen.

When we reuse it there, we will want to track the analytics according to the flow that the user is following.

This renames the analytics used on that screen, and adds a `flow` parameter, to indicate whether a payment method event was triggered while following the flow from the Order details screen, or the Simple Payments flow.