6782 HostedPaymentMethodsView for toolbar items
The PaymentMethodsView is presented in both the Simple Payments flow, in a SwiftUI navigation view, and from Order details, in a UIHostingController.

In the Simple Payments flow it is part of a sequence of modally presented actions, so the correct button is a back navigation button, which the view gets for free because of its presentation.

When presented modally from Order Details, it is not part of a wider flow, and is the only view in the navstack. This adds a cancel button when shown in Order Details, by wrapping the view in HostedPaymentMethodsView.