7044 Remove PaymentGatewayAccounts on OrderDetails
PaymentGatewayAccounts were used to determine whether or not to show a card payment button.

The card payment button has been replaced by a generic Collect Payment button, which we show to any merchant.

Therefore, PaymentGatewayAccounts are no longer needed on Order Details, and this change removes the code associated with fetching and refreshing them.