Remove onboarding data refresh from Order Details
In Ducks in a Row, we replaced the Card Payment button on Order Details with a Collect Payment button that allowed for multiple payment methods to be shown and chosen from.

The result of this is that the OrderDetailsViewModel no longer needs to present the CPP onboarding, and this change removes the obsolete refreshing.