Retrieve order after marking as paid
When used from Order Details, the PaymentMethodsView would be dismissed to reveal the Order Details, and after marking an order as paid, the Collect Order button was still visible.

This was due to the order’s `datePaid` property still being nil, because the updated order had not been fetched, only the status had been updated.

This change refreshes the Order after we successfully update the status, to make sure that the Order Details screen shows fresh data and does not allow the order to be collected twice.