7157 Track cancelation on card reader modals.
Previously, we used `onFailure` and `onCompletion` which were each triggered for some cancelation paths. The `onFailure` closure had an error that could be used to check whether it was a cancellation, but `onCompletion` did not, and was handled as a success.

This meant that the `canceled` Tracks event was logged along with the `completed` event or `failed` event.

Cancelation is not a failure, and is not completion either.

This change adds a specific handler for cancelation, and translates the `.flowCanceledByUser` errors to a cancel event, so that Tracks events are correctly logged.