Removed save-tracking side-effects from canCommit
Previously any check for wether the form was ready to be committed would also save the tracking provider, this happen at least twice for each use of the screen and resulted in a tracking save. Removing the save from canCommit means we only save now when the user taps the "add tracking" button.