Enable scrolling on the PaymentMethod screen
Previously, there was only ever a maximum of three rows on the PaymentMethod screen, and they were all standard height. This meant there was no situation when that screen would need to scroll.

Recently, we added a feature announcement card, which does not completely fit on smaller devices in landscape.

This adds a scrollview to the PaymentMethod screen to enable users to see the full content of that screen, including when they are in landscape.