Move cash alert so that child view alerts will work
A child view in a heirarchy is not able to show an alert, if there’s an alert modifier on its parent.

We will add an alert to the FeatureAnnouncementCard in #7252, so the alert for Cash payment on the Payment Method screen must be moved up to allow for that.

Readability is also improved by keeping the alert definition close to where it is triggered.