Remind (14 days) later on Feature A10t Cards
Feature Announcement Cards can be dismissed by the user. Some users may wish to temporarily get a card out of the way, but be interested in acting on it in future.

This adds a choice for the user when dismissing a card: “Don’t show again”, or “Remind me later”.

A user who chooses “Remind me later” will be shown the card again after 14 days.