7502 Skip on Pending Requirements IPP Onboarding
The onboarding state for `Stripe account has pending requirements` informs the merchant that there's additional information needed on their account, but that they have some time to complete that information before it becomes a requirement. The deadline is also shown.

Previously, it did not have a button, so it prevented merchants taking payments even though the deadline had not passed.

This change adds a skip button so that the merchant can still take payments. Once the deadline is past, the overdue requirements screen is shown instead, which does not have a skip button.