7540 Display maximum one notice at a time
Previously, the PermanentNoticePresenter would display a new notice every time `presentNotice` was called, losing it’s reference to any existing notice, but leaving it in the view hierarchy.

This meant that in the In-Person Payments onboarding flow, the “Continue Setup” notice would be displayed once for every failure that was shown, and only the top one removed by `dismiss` when the `.completed` step was reached.

This change makes `presentNotice` do nothing if a notice is already presented, so in the In-Person Payments scenario, there will only ever be one presented and dismissed.

The PermanentNoticePresenter is not currently used elsewhere.