Set `skip_prechecks:true` in CI release builds
This will result in CI relying on the CocoaPods cache it just downloaded
instead of reinstalling all pods from scratch.

It should save around a minute of CI time.
See this build for example:
https://buildkite.com/automattic/woocommerce-ios/builds/7674#01830b20-18fd-4cef-bedd-5d494da79152/1163-1249

Admittedly, one minute over a total of 26 is not that much.
Still, we pay CI by the minute and, pragmatically, we shouldn't run
unnecessary operations if we can avoid it.