7677 Cancel discovery when user taps cancel
When a reader is found, the user has the option to Connect, Keep Searching, or Cancel.

Previously, if they cancelled the reader connection, the prompt was dismissed, but the SDK remained in the discoverReaders mode indefinitely.

This resulted in the user being able to start a new connection attempt.

This change calls `cancel` on the Cancellable that Stripe Terminal gives us, which results in the SDK being usable again.