7686 Match presentationStyle on connection modals
Previously, CardPresentPaymentsModalViewController modals used the FancyAlertController custom presentation, while SeveralReadersFoundViewController used the system `overFullScreen` method, combined with a view background colour set in the nib.

This resulted in differences in the dimming view (as the FancyAlertController used alpha: 0.5, while the background view used 0.3-ish).

It also resulted in some discrepancies around animating the transitions between view controllers. These are not fully resolved here, but they are improved.

The presentation and transition style for `overFullScreen` matches FancyAlertView, so it seems a better choice. This updates both views to use that style with a 50% opacity background.